requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
Pillow
torch
torchvision